% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cluster.R
\name{find_cluster}
\alias{find_cluster}
\title{GPS Cluster Identification}
\usage{
find_cluster(ind = "Cluster Locations", data = data, time_thres = 1,
  dist_thres = 0.05, size = 1)
}
\arguments{
\item{ind}{Text for title of figure of cluster locations produced by 'find_cluster'. Defaults to "Cluster Locations"}

\item{data}{A data.frame that contains 'date', 'Time', 'lat', and 'long' columns}

\item{time_thres}{Time, measured in days, between GPS locations to be considered as part of a cluster (i.e., 1 = 24 hours)}

\item{dist_thres}{Distance, measured in kilometers, between GPS points to be considered in a cluster (i.e., 0.05 = 50 meters)}
}
\description{
This function identifies potential predation sites ("clusters") by using GPS locations from GPS collared carnivores and user defined thresholds.
}
\examples{
# Identify clusters formed by GPS locations that occur within 1 day of eachother 
# and within 100 meters of eachother.

find_cluster(ind = "Cluster Locations", data = data, time_thres = 1, dist_thres = 0.1, size = 1)
}
\keyword{Carnivore}
\keyword{Clusters}
\keyword{GPS}
\keyword{Predation}
