% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRwolves.R
\name{IRwolves}
\alias{IRwolves}
\title{Isle Royale Wolf GPS Cluster Identification}
\usage{
IRwolves(DownloadDate = DownloadDate, CollarID = CollarID,
  AnimalID = AnimalID, NextCluster = NextCluster,
  ClusterDate = ClusterDate, time_thres = time_thres,
  dist_thres = dist_thres)
}
\arguments{
\item{DownloadDate}{The date and folder title in which the raw GPS data are housed, should be formated as DDMONYYYY (i.e., 14MAY2019)}

\item{CollarID}{The individual identifier for each collar preceeded by "Collar" (i.e., Collar32712)}

\item{AnimalID}{The individual identifier for each wolf (i.e., W001F) indicating the wolf number and sex}

\item{NextCluster}{The next number to be assinged for cluster labeling for independent identification, reference Download_Dates.xlsx}

\item{ClusterDate}{The oldest date to be used for the period to identify clusters within (format = "YYYY-MM-DD", i.e., "2019-05-14")}

\item{time_thres}{Time, measured in days, between GPS locations to be considered as part of a cluster (i.e., 1 = 24 hours)}

\item{dist_thres}{Distance, measured in kilometers, between GPS points to be considered in a cluster (i.e., 0.05 = 50 meters)}
}
\description{
This function formats GPS data from Telonics and Vectoronic Collars used in Isle Royale National Park and then identifies potential predation sites ("clusters") from the formated GPS locations and user defined thresholds.
}
\examples{
# Identify clusters formed by GPS locations that occur within 1 day of eachother
# and within 50 meters of eachother.

IRwolves <- function(DownloadDate = "14MAY2019", CollarID = "Collar32712", AnimalID = "W001F", NextCluster = 001, ClusterDate = "2019-05-07", time_thres = 1, dist_thres = 0.05)
}
\keyword{Clusters}
\keyword{GPS}
\keyword{Isle}
\keyword{Predation}
\keyword{Royale}
\keyword{Wolf}
